requirements.txt
# Web & API
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-SocketIO==5.3.6
python-socketio==5.11.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.1
gunicorn==21.2.0

# Socket.IO workers (pick one in production)
eventlet==0.33.3
gevent  # optional fallback worker

# Data stores & messaging
pymongo==4.6.1          # also provides bson/gridfs
pika==1.3.2
requests==2.31.0
urllib3>=2.0.0
pytz==2024.1

# Audio & data prep
numpy>=1.24.3
pandas>=2.0.0
librosa==0.10.1
soundfile==0.12.1
sounddevice>=0.4.6
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.3.0

# ML/analysis stack
# torch>=2.0.0
# torchvision>=0.15.0
